+++ odoo/local-src/bso_dealsheet/__manifest__.py
@@ -4,7 +4,13 @@
     'description': 'BSO Dealsheet Costs, Margins & Validation Process',
     'author': 'BSO',
     'website': 'https://www.bsonetwork.com',
-    'depends': ['sale', 'product_bundle', 'specific_sale', 'bso_backbone'],
+    'depends': [
+        'sale',
+        'product_bundle',
+        'specific_sale',
+        'bso_backbone',
+        'purchase'
+    ],
     'data': [
         'security/ir.model.access.csv',
         'views/sale_dealsheet.xml',
@@ -14,6 +20,7 @@
         'views/sale_dealsheet_wizard_refuse.xml',
         'views/sale_dealsheet_wizard_request.xml',
         'views/sale_order.xml',
+        'wizard/wiz_sale_dealsheet_source.xml',
     ],
     'installable': True,
     'auto_install': False,