+++ odoo/local-src/bso_dealsheet/models/__init__.py
@@ -5,3 +5,4 @@ from . import sale_dealsheet_wizard_confirm
 from . import sale_dealsheet_wizard_refuse
 from . import sale_dealsheet_wizard_request
 from . import sale_order
+from . import purchase_order