+++ odoo/local-src/bso_dealsheet/models/sale_dealsheet.py
@@ -358,6 +358,20 @@ class SaleDealsheet(models.Model):
             'date_validated': fields.Datetime.now()
         })
 
+    def _load_win_action(self, act, **kw):
+        act = self.env.ref(act)
+        action = act.copy_data()[0]
+        action['context'] = self._context.copy()
+        action.update(kw)
+        return action
+
+    @api.multi
+    def action_procure(self):
+        action = self._load_win_action(
+            'bso_dealsheet.act_wiz_sale_dealsheet_source')
+        action['context']['default_dealsheet_id'] = self.id
+        return action
+
     # TOOLS
 
     @api.model